+++ .idea/misc.xml
@@ -10,10 +10,12 @@
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/activity_splash.xml" value="0.22813688212927757" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_cale.xml" value="0.1" />
+        <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_graph.xml" value="0.17603850050658562" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_market.xml" value="0.1" />
+        <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_tab_timer.xml" value="0.1" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_til.xml" value="0.1" />
-        <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_timer.xml" value="0.25" />
+        <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_timer.xml" value="0.1" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/layout/fragment_tree.xml" value="0.1" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/menu/botton_nav_menu.xml" value="0.1" />
         <entry key="..\:/Users/Admin/AndroidStudioProjects/HadTo3/app/src/main/res/menu/fragment_home.xml" value="0.1" />
