+++ app/src/main/java/com/example/hadto/timerTapbar/tabGraphFragment.kt
@@ -1,0 +1,24 @@
+package com.example.hadto.timerTapbar
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.hadto.R
+
+
+class tabGraphFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+
+
+        return inflater.inflate(R.layout.fragment_graph, container, false)
+    }
+
+}
