+++ app/src/main/res/layout/fragment_graph.xml
@@ -3,12 +3,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".timerTapbar.GraphFragment">
+    tools:context=".timerTapbar.tabGraphFragment">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:layout_height="match_parent">
 
+
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
 </FrameLayout>
