+++ app/src/main/res/layout/fragment_tab_timer.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".timerTapbar.tabTimerFragment">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <View
+            android:id="@+id/circleView"
+            android:layout_width="300dp"
+            android:layout_height="300dp"
+            android:layout_marginTop="90dp"
+            android:background="@drawable/round_bg"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
+
+        <TextView
+            android:id="@+id/timerNum"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="5dp"
+            android:text="@string/timerZeroText"
+            android:textSize="70sp"
+            app:layout_constraintBottom_toBottomOf="@+id/circleView"
+            app:layout_constraintEnd_toEndOf="@+id/circleView"
+            app:layout_constraintStart_toStartOf="@+id/circleView"
+            app:layout_constraintTop_toTopOf="@+id/circleView"
+            android:textColor="@color/white"/>
+
+        <Button
+            android:id="@+id/stopBtn"
+            android:layout_width="160dp"
+            android:layout_height="70dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/circleView"
+            android:layout_marginStart="30dp"
+            android:layout_marginTop="15dp"
+            android:textSize="35sp"
+            android:text="@string/restText"/>
+
+        <Button
+            android:id="@+id/startBtn"
+            android:layout_width="160dp"
+            android:layout_height="70dp"
+            app:layout_constraintStart_toEndOf="@+id/stopBtn"
+            app:layout_constraintTop_toTopOf="@+id/stopBtn"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:text="@string/startText"
+            android:textSize="35sp" />
+
+        <Button
+            android:id="@+id/cancelBtn"
+            android:layout_width="180dp"
+            android:layout_height="70dp"
+            app:layout_constraintTop_toBottomOf="@+id/stopBtn"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:text="@string/cancelText"
+            android:textSize="35sp"
+            android:layout_marginTop="8dp"/>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</FrameLayout>
