+++ app/src/main/res/layout/fragment_timer.xml
@@ -3,51 +3,36 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:id="@+id/tabContent">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constrainLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
+
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tabs_main"
+            app:layout_constraintTop_toTopOf="parent"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"/>
+
         <TextView
             android:id="@+id/dateText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="20sp"
+            android:layout_marginStart="10sp"
             android:layout_marginTop="20dp"
             android:fontFamily="@font/abel"
             android:textColor="@color/black"
-            android:text="2022-01-01"
-            android:textSize="50sp"
-            android:background="@color/dateTextBackground"
+            android:textSize="42sp"
             android:textStyle="bold"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toBottomOf="@+id/tabs_main" />
 
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/tabs_main"
-            app:layout_constraintTop_toBottomOf="@+id/dateText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <com.google.android.material.tabs.TabItem
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/timerTapText"/>
-            <com.google.android.material.tabs.TabItem
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/graphTapText"/>
-
-            <com.google.android.material.tabs.TabItem
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/treeTapText"/>
-
-
-        </com.google.android.material.tabs.TabLayout>
-        <androidx.viewpager.widget.ViewPager
-            android:id="@+id/viewpager_main"
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/viewPager2"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:layout_constraintStart_toStartOf="parent"
@@ -55,60 +40,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintTop_toTopOf="parent"/>
 
-        <View
-            android:id="@+id/circleView"
-            android:layout_width="300dp"
-            android:layout_height="300dp"
-            android:layout_marginTop="20dp"
-            android:background="@drawable/round_bg"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tabs_main"
-            app:layout_constraintStart_toStartOf="parent" />
 
-        <TextView
-            android:id="@+id/timerNum"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="5dp"
-            android:text="@string/timerZeroText"
-            android:textSize="70sp"
-            app:layout_constraintBottom_toBottomOf="@+id/circleView"
-            app:layout_constraintEnd_toEndOf="@+id/circleView"
-            app:layout_constraintStart_toStartOf="@+id/circleView"
-            app:layout_constraintTop_toTopOf="@+id/circleView"
-            android:textColor="@color/white"/>
-
-        <Button
-            android:id="@+id/stopBtn"
-            android:layout_width="160dp"
-            android:layout_height="70dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/circleView"
-            android:layout_marginStart="30dp"
-            android:layout_marginTop="23dp"
-            android:textSize="35sp"
-            android:text="@string/restText"/>
-
-        <Button
-            android:id="@+id/startBtn"
-            android:layout_width="160dp"
-            android:layout_height="70dp"
-            app:layout_constraintStart_toEndOf="@+id/stopBtn"
-            app:layout_constraintTop_toTopOf="@+id/stopBtn"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:text="@string/startText"
-            android:textSize="35sp" />
-
-        <Button
-            android:id="@+id/cancelBtn"
-            android:layout_width="180dp"
-            android:layout_height="80dp"
-            app:layout_constraintTop_toBottomOf="@+id/stopBtn"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:text="@string/cancelText"
-            android:textSize="40sp"
-            android:layout_marginTop="8dp"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
